refactor: move hex & improve css
